+++ src/common/utils/gitUtils.ts
@@ -181,7 +181,7 @@ export async function computeCommitsSummary(checkOnly, pullRequestInfo: any) {
 
   // Add Flow diff in Markdown
   let flowDiffMarkdown: any = {};
-  if (checkOnly || GitProvider.isDeployBeforeMerge() && !(process.env?.SFDX_DISABLE_FLOW_DIFF === "true")) {
+  if ((checkOnly || GitProvider.isDeployBeforeMerge()) && !(process.env?.SFDX_DISABLE_FLOW_DIFF === "true")) {
     const flowList: string[] = [];
     for (const logResult of logResults) {
       const updatedFiles = await getCommitUpdatedFiles(logResult.hash);
@@ -240,7 +240,7 @@ export async function buildCheckDeployCommitSummary() {
   }
 }
 
-export async function handlePostDeploymentNotifications(flags, targetUsername: any, quickDeploy: any, delta: boolean, debugMode: boolean) {
+export async function handlePostDeploymentNotifications(flags, targetUsername: any, quickDeploy: any, delta: boolean, debugMode: boolean, additionalMessage = "") {
   const pullRequestInfo = await GitProvider.getPullRequestInfo();
   const attachments: MessageAttachment[] = [];
   try {
@@ -270,6 +270,9 @@ export async function handlePostDeploymentNotifications(flags, targetUsername: a
     : delta
       ? ' (🌙 delta deployment)'
       : ' (🌕 full deployment)';
+  if (additionalMessage) {
+    notifMessage += '\n\n' + additionalMessage + "\n\n"
+  }
 
   const notifButtons = await getNotificationButtons();
   if (pullRequestInfo) {